refactor: вынести функцию фильтрации фильмов
